Fix: Class "EmailMessage" has 28 methods, which is greater than 20 authorized. Split it into smaller classes.
